feat: update VSCode configurations and add new event handling functionality
