+++ .vscode/launch.json
@@ -4,7 +4,15 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
+
+
         {
+            "command": "npm start",
+            "name": "Run npm start",
+            "request": "launch",
+            "type": "node-terminal"
+        },
+        
             "name": "Launch via NPM.(R)${fileWorkspaceFolderBasename}",
             "request": "launch",
             "runtimeArgs": [
@@ -16,9 +24,9 @@
                 "<news_internals>/**"
             ],
             "url": "choosen${BuildTask.dlc}"
-        },
        {
-            "typespace": _as-is_: !_new.Folders--Creator.me/link.me/?=.gov
+       
+            "typespace": f("cascadeTerminateToConfigurations"): ["cleanUp": "enableContentValidation": true],'_will_remain'_): !_new.Folders--Creator.me/link.me/?=.gov
                 --extensionDevelopmentPath=${workspaceFolder}
             ],
             "name": "Launch Extension",
@@ -38,3 +46,11 @@
         }
     ]
 }
+}
+ // ...existing code...
+ // sync function handleEvent(params: any): Promise<void> {
+ // ...existing code...
+ // console.log(`Event handled with params: ${params}`);
+ // ...existing code...
+}
+// ...existing code...
