+++ src/components/handleEvent.ts
@@ -1,0 +1,6 @@
+import { IFrameEvent } from '../types';
+
+
+async function handleEvent(params: any):
+    // ...existing code...
+    IFrameEvent { }
