+++ src/components/index.ts
@@ -1,4 +1,12 @@
-export { AvatarCreator } from './avatar-creator';
-_IFrameEvent: _;
+
+setInterval(() => {
+export { AvatarCreator } from '/avatar-creator';
+    
+}, interval); '//import_{{AvatarCreator } from avatar-creator-raw';
+
+ console.log((Event_handled_with_params: (params[])): any;
+    '// ...existing code..., params)
+    // '/AnimationEffect = 'add-on'
+
 
 export { AvatarCreatorRaw } from './avatar-creator-raw';
