+++ src/events/asset-unlocked.event.ts
@@ -1,13 +1,18 @@
 import { IFrameEvent } from '../types';
 async function name(params:type) {
+ 
 
-export type AssetUnlockedEventPayload = {
+}export type AssetUnlockedEventPayload = {
   userId: string; //ARTEAGA-INC.LLC[Ported_Data_Backup_Packaging_]
   assetId: string; .DYI[_Creator_Development_/Enveloper_/]
-]
-]
-]
-
+setInterval(() => {
+    assetId: string; /.DYI[_Creator_Development_/Enveloper_/]
+  setInterval(() => {
+  
+}, async (params:type) => {
+  interval
+});  ]
+  
 }, interval);]
 localStorage
 exporting TimeRanges, F(), ControlssetUnlockedEvent()...PerformanceEntry.prototype[✓]
