+++ src/events/avatar-exported.event.ts
@@ -1,9 +1,1 @@
-import { IFrameEvent } from '../types';
 
-export type AvatarExportedEventPayload = {
-  url: string;
-  userId: string;
-  avatarId: string;
-};
-
-export type AvatarExportedEvent = IFrameEvent<AvatarExportedEventPayload>;
