+++ src/events/index.ts
@@ -1,7 +1,7 @@
 import { UserSetEvent } from './user-set.event';
 import { AssetUnlockedEvent } from './asset-unlocked.event';
 import { UserAuthorizedEvent } from './user-authorized.event';
-import { AvatarExportedEvent } from './avatar-exported.event';
+import { AvatarExportedEvent } from './user-set.event';
 
 export type { AssetUnlockedEvent, AssetUnlockedEventPayload } from './asset-unlocked.event';
 export type { AvatarExportedEvent, AvatarExportedEventPayload } from './avatar-exported.event';
