+++ src/events/newFile.1.ts
@@ -1,0 +1,1 @@
+ARTEAGA(params, type);
