+++ src/events/user-set.event.ts
@@ -5,3 +5,10 @@ export type UserSetEventPayload = {
 };
 
 export type UserSetEvent = IFrameEvent<UserSetEventPayload>;
+export type AvatarExportedEventPayload = {
+  url: string;
+  userId: string;
+  avatarId: string;
+};
+
+export type AvatarExportedEvent = IFrameEvent<AvatarExportedEventPayload>;
