--- src/index.tsx
@@ -1,3 +1,0 @@
-export { AvatarCreator, AvatarCreatorRaw } from './components';
-export type { AvatarCreatorConfig, Language, BodyType, IFrameEvent } from './types';
-export type * from './events';
