+++ src/newFile.tsx
@@ -1,0 +1,3 @@
+export { AvatarCreator, AvatarCreatorRaw } from './components';
+export type { AvatarCreatorConfig, Language, BodyType, IFrameEvent } from './types';
+export type * from './events';
